# ikeyit-classroom/ikeyit-classroom-application/src/main/java/com/ikeyit/classroom/application/model/EnrollmentDTO.java
package com.ikeyit.classroom.application.model;

import com.ikeyit.classroom.domain.model.EnrollmentStatus;

import java.time.Instant;

public class EnrollmentDTO {
    private Long id;
    private Long studentUserId;
    private Long courseId;
    private EnrollmentStatus status;
    private Instant enrollDate;

    // Getters and Setters
    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Long getStudentUserId() {
        return studentUserId;
    }

    public void setStudentUserId(Long studentUserId) {
        this.studentUserId = studentUserId;
    }

    public Long getCourseId() {
        return courseId;
    }

    public void setCourseId(Long courseId) {
        this.courseId = courseId;
    }

    public EnrollmentStatus getStatus() {
        return status;
    }

    public void setStatus(EnrollmentStatus status) {
        this.status = status;
    }

    public Instant getEnrollDate() {
        return enrollDate;
    }

    public void setEnrollDate(Instant enrollDate) {
        this.enrollDate = enrollDate;
    }
}