# ikeyit-classroom/ikeyit-classroom-application/src/main/java/com/ikeyit/classroom/application/model/PublishCourseCMD.java
package com.ikeyit.classroom.application.model;

public class PublishCourseCMD {
    private Long operatorId;
    private Long courseId;

    public Long getOperatorId() {
        return operatorId;
    }

    public void setOperatorId(Long operatorId) {
        this.operatorId = operatorId;
    }

    public Long getCourseId() {
        return courseId;
    }

    public void setCourseId(Long courseId) {
        this.courseId = courseId;
    }
}